+++ Changelog.rst
@@ -5,6 +5,8 @@ Version NEXTVERSION
 
 * Fix bug in `cfdm.read` that prevented some OPeNDAP URLS being read
   (https://github.com/NCAS-CMS/cfdm/issues/406)
+* Fix bug in `cfdm.write` when writing numpy scalars as Zarr
+  attributes (https://github.com/NCAS-CMS/cfdm/issues/408)
 
 Version 1.13.1.0
 ----------------